Some fixes for double/float processing

+++ hecuba/hdict.py
@@ -469,24 +469,24 @@ class StorageDict(dict, IStorage):
                 col_types = map(lambda x: IStorage._conversions[x.__class__.__name__], val)
                 spec_col_types = map(lambda x: x[1], self._columns)
                 for idx, value in enumerate(spec_col_types):
-                    if value == 'float':
-                        spec_col_types[idx] = 'double'
+                    if value == 'double':
+                        spec_col_types[idx] = 'float'
             else:
                 col_types = IStorage._conversions[val.__class__.__name__]
                 spec_col_types = map(lambda x: x[1], self._columns)[0]
-                if spec_col_types == 'float':
-                    spec_col_types = 'double'
+                if spec_col_types == 'double':
+                    spec_col_types = 'float'
             if isinstance(key, Iterable) and not isinstance(key, str):
                 key_types = map(lambda x: IStorage._conversions[x.__class__.__name__], key)
                 spec_key_types = map(lambda x: x[1], self._primary_keys)
                 for idx, value in enumerate(spec_key_types):
-                    if value == 'float':
-                        spec_key_types[idx] = 'double'
+                    if value == 'double':
+                        spec_key_types[idx] = 'float'
             else:
                 key_types = IStorage._conversions[key.__class__.__name__]
                 spec_key_types = map(lambda x: x[1], self._primary_keys)[0]
-                if spec_key_types == 'float':
-                    spec_key_types = 'double'
+                if spec_key_types == 'double':
+                    spec_key_types = 'float'
             if (col_types == spec_col_types):
                 if(key_types == spec_key_types):
                     if not self._is_persistent:
